requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
ultralytics>=8.0.0  # YOLOv8

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
scikit-image>=0.21.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data Versioning
dvc>=3.0.0
dvc-s3>=3.0.0

# MLOps
mlflow>=2.7.0
wandb>=0.15.0

# API & Serving
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Monitoring & Metrics
prometheus-client>=0.17.0
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0
